Sync with upstream

+++ CRM/Core/I18n/Schema.php
@@ -1,9 +1,9 @@
 <?php
 /*
  +--------------------------------------------------------------------+
- | CiviCRM version 4.5                                                |
+ | CiviCRM version 5                                                  |
  +--------------------------------------------------------------------+
- | Copyright CiviCRM LLC (c) 2004-2014                                |
+ | Copyright CiviCRM LLC (c) 2004-2018                                |
  +--------------------------------------------------------------------+
  | This file is a part of CiviCRM.                                    |
  |                                                                    |
@@ -23,23 +23,19 @@
  | GNU Affero General Public License or the licensing of CiviCRM,     |
  | see the CiviCRM license FAQ at http://civicrm.org/licensing        |
  +--------------------------------------------------------------------+
-*/
+ */
 
 /**
  *
  * @package CRM
- * @copyright CiviCRM LLC (c) 2004-2014
- * $Id$
- *
+ * @copyright CiviCRM LLC (c) 2004-2018
  */
 class CRM_Core_I18n_Schema {
 
   /**
    * Drop all views (for use by CRM_Core_DAO::dropAllTables() mostly).
-   *
-   * @return void
    */
-  static function dropAllViews() {
+  public static function dropAllViews() {
     $domain = new CRM_Core_DAO_Domain();
     $domain->find(TRUE);
     if (!$domain->locales) {
@@ -60,11 +56,10 @@ class CRM_Core_I18n_Schema {
    * Switch database from single-lang to multi (by adding
    * the first language and dropping the original columns).
    *
-   * @param $locale string  the first locale to create (migrate to)
-   *
-   * @return void
+   * @param string $locale
+   *   the first locale to create (migrate to).
    */
-  static function makeMultilingual($locale) {
+  public static function makeMultilingual($locale) {
     $domain = new CRM_Core_DAO_Domain();
     $domain->find(TRUE);
 
@@ -83,14 +78,18 @@ class CRM_Core_I18n_Schema {
       // drop old indices
       if (isset($indices[$table])) {
         foreach ($indices[$table] as $index) {
-          $queries[] = "DROP INDEX {$index['name']} ON {$table}";
+          if (CRM_Core_BAO_SchemaHandler::checkIfIndexExists($table, $index['name'])) {
+            $queries[] = "DROP INDEX {$index['name']} ON {$table}";
+          }
         }
       }
       // deal with columns
       foreach ($hash as $column => $type) {
         $queries[] = "ALTER TABLE {$table} ADD {$column}_{$locale} {$type}";
-        $queries[] = "UPDATE {$table} SET {$column}_{$locale} = {$column}";
-        $queries[] = "ALTER TABLE {$table} DROP {$column}";
+        if (CRM_Core_BAO_SchemaHandler::checkIfFieldExists($table, $column)) {
+          $queries[] = "UPDATE {$table} SET {$column}_{$locale} = {$column}";
+          $queries[] = "ALTER TABLE {$table} DROP {$column}";
+        }
       }
 
       // add view
@@ -108,18 +107,20 @@ class CRM_Core_I18n_Schema {
     // update civicrm_domain.locales
     $domain->locales = $locale;
     $domain->save();
+
+    // CRM-21627 Updates the $dbLocale
+    CRM_Core_BAO_ConfigSetting::applyLocale(Civi::settings($domain->id), $domain->locales);
   }
 
   /**
    * Switch database from multi-lang back to single (by dropping
    * additional columns and views and retaining only the selected locale).
    *
-   * @param $retain string  the locale to retain
-   *
-   * @return void
+   * @param string $retain
+   *   the locale to retain.
    */
-  static function makeSinglelingual($retain) {
-    $domain = new CRM_Core_DAO_Domain;
+  public static function makeSinglelingual($retain) {
+    $domain = new CRM_Core_DAO_Domain();
     $domain->find(TRUE);
     $locales = explode(CRM_Core_DAO::VALUE_SEPARATOR, $domain->locales);
 
@@ -129,8 +130,8 @@ class CRM_Core_I18n_Schema {
     }
 
     // lets drop all triggers first
-    $logging = new CRM_Logging_Schema;
-    $logging->dropTriggers( );
+    $logging = new CRM_Logging_Schema();
+    $logging->dropTriggers();
 
     // turn subsequent tables singlelingual
     $tables = CRM_Core_I18n_SchemaStructure::tables();
@@ -147,27 +148,28 @@ class CRM_Core_I18n_Schema {
     $dbLocale = '';
 
     // now lets rebuild all triggers
-    CRM_Core_DAO::triggerRebuild( );
+    CRM_Core_DAO::triggerRebuild();
   }
 
   /**
    * Switch a given table from multi-lang to single (by retaining only the selected locale).
    *
-   * @param $retain string  the locale to retain
-   * @param $table  string  the table containing the column
-   * @param $class  string  schema structure class to use to recreate indices
+   * @param string $retain
+   *   the locale to retain.
+   * @param string $table
+   *   the table containing the column.
+   * @param string $class
+   *   schema structure class to use to recreate indices.
    *
    * @param array $triggers
-   *
-   * @return void
    */
-  static function makeSinglelingualTable(
+  public static function makeSinglelingualTable(
     $retain,
     $table,
     $class = 'CRM_Core_I18n_SchemaStructure',
     $triggers = array()
   ) {
-    $domain = new CRM_Core_DAO_Domain;
+    $domain = new CRM_Core_DAO_Domain();
     $domain->find(TRUE);
     $locales = explode(CRM_Core_DAO::VALUE_SEPARATOR, $domain->locales);
 
@@ -189,25 +191,27 @@ class CRM_Core_I18n_Schema {
       }
     }
 
+    $dao = new CRM_Core_DAO();
     // deal with columns
     foreach ($columns[$table] as $column => $type) {
-      $queries[] = "ALTER TABLE {$table} ADD {$column} {$type}";
-      $queries[] = "UPDATE {$table} SET {$column} = {$column}_{$retain}";
+      $queries[] = "ALTER TABLE {$table} CHANGE `{$column}_{$retain}` `{$column}` {$type}";
       foreach ($locales as $loc) {
-        $dropQueries[] = "ALTER TABLE {$table} DROP {$column}_{$loc}";
+        if (strcmp($loc, $retain) !== 0) {
+          $dropQueries[] = "ALTER TABLE {$table} DROP {$column}_{$loc}";
+        }
       }
     }
 
     // drop views
     foreach ($locales as $loc) {
-      $queries[] = "DROP VIEW {$table}_{$loc}";
+      $queries[] = "DROP VIEW IF EXISTS {$table}_{$loc}";
     }
 
     // add original indices
     $queries = array_merge($queries, self::createIndexQueries(NULL, $table));
 
     // execute the queries without i18n rewriting
-    $dao = new CRM_Core_DAO;
+    $dao = new CRM_Core_DAO();
     foreach ($queries as $query) {
       $dao->query($query, FALSE);
     }
@@ -216,31 +220,31 @@ class CRM_Core_I18n_Schema {
       $dao->query($query, FALSE);
     }
 
-    if ( !empty($triggers)) {
+    if (!empty($triggers)) {
       if (CRM_Core_Config::isUpgradeMode()) {
-      foreach ($triggers as $triggerInfo) {
-        $when = $triggerInfo['when'];
-        $event = $triggerInfo['event'];
-        $triggerName = "{$table}_{$when}_{$event}";
-        CRM_Core_DAO::executeQuery("DROP TRIGGER IF EXISTS {$triggerName}");
+        foreach ($triggers as $triggerInfo) {
+          $when = $triggerInfo['when'];
+          $event = $triggerInfo['event'];
+          $triggerName = "{$table}_{$when}_{$event}";
+          CRM_Core_DAO::executeQuery("DROP TRIGGER IF EXISTS {$triggerName}");
+        }
       }
-    }
 
-    // invoke the meta trigger creation call
-    CRM_Core_DAO::triggerRebuild($table);
-  }
+      // invoke the meta trigger creation call
+      CRM_Core_DAO::triggerRebuild($table);
+    }
   }
 
   /**
    * Add a new locale to a multi-lang db, setting
    * its values to the current default locale.
    *
-   * @param $locale string  the new locale to add
-   * @param $source string  the locale to copy from
-   *
-   * @return void
+   * @param string $locale
+   *   the new locale to add.
+   * @param string $source
+   *   the locale to copy from.
    */
-  static function addLocale($locale, $source) {
+  public static function addLocale($locale, $source) {
     // get the current supported locales
     $domain = new CRM_Core_DAO_Domain();
     $domain->find(TRUE);
@@ -261,7 +265,7 @@ class CRM_Core_I18n_Schema {
       // add new columns
       foreach ($hash as $column => $type) {
         // CRM-7854: skip existing columns
-        if (CRM_Core_DAO::checkFieldExists($table, "{$column}_{$locale}", FALSE)) {
+        if (CRM_Core_BAO_SchemaHandler::checkIfFieldExists($table, "{$column}_{$locale}", FALSE)) {
           continue;
         }
         $queries[] = "ALTER TABLE {$table} ADD {$column}_{$locale} {$type}";
@@ -292,12 +296,14 @@ class CRM_Core_I18n_Schema {
   /**
    * Rebuild multilingual indices, views and triggers (useful for upgrades)
    *
-   * @param $locales array  locales to be rebuilt
-   * @param $version string version of schema structure to use
-   *
-   * @return void
+   * @param array $locales
+   *   locales to be rebuilt.
+   * @param string $version
+   *   version of schema structure to use.
+   * @param bool $isUpgradeMode
+   *   Are we upgrading our database
    */
-  static function rebuildMultilingualSchema($locales, $version = NULL) {
+  public static function rebuildMultilingualSchema($locales, $version = NULL, $isUpgradeMode = FALSE) {
     if ($version) {
       $latest = self::getLatestSchema($version);
       require_once "CRM/Core/I18n/SchemaStructure_{$latest}.php";
@@ -307,9 +313,9 @@ class CRM_Core_I18n_Schema {
       $class = 'CRM_Core_I18n_SchemaStructure';
     }
     $indices =& $class::indices();
-    $tables  =& $class::tables();
+    $tables =& $class::tables();
     $queries = array();
-    $dao = new CRM_Core_DAO;
+    $dao = new CRM_Core_DAO();
 
     // get all of the already existing indices
     $existing = array();
@@ -338,7 +344,7 @@ class CRM_Core_I18n_Schema {
     // rebuild views
     foreach ($locales as $locale) {
       foreach ($tables as $table) {
-        $queries[] = self::createViewQuery($locale, $table, $dao, $class);
+        $queries[] = self::createViewQuery($locale, $table, $dao, $class, $isUpgradeMode);
       }
     }
 
@@ -356,15 +362,20 @@ class CRM_Core_I18n_Schema {
   /**
    * Rewrite SQL query to use views to access tables with localized columns.
    *
-   * @param $query string  the query for rewrite
+   * @param string $query
+   *   the query for rewrite.
    *
-   * @return string        the rewritten query
+   * @return string
+   *   the rewritten query
    */
-  static function rewriteQuery($query) {
+  public static function rewriteQuery($query) {
     global $dbLocale;
     $tables = self::schemaStructureTables();
     foreach ($tables as $table) {
-      $query = preg_replace("/([^'\"])({$table})([^_'\"])/", "\\1\\2{$dbLocale}\\3", $query);
+      // CRM-19093
+      // should match the civicrm table name such as: civicrm_event
+      // but must not match the table name if it's a substring of another table: civicrm_events_in_cart
+      $query = preg_replace("/([^'\"])({$table})(\z|[^a-z_'\"])/", "\\1\\2{$dbLocale}\\3", $query);
     }
     // uncomment the below to rewrite the civicrm_value_* queries
     // $query = preg_replace("/(civicrm_value_[a-z0-9_]+_\d+)([^_])/", "\\1{$dbLocale}\\2", $query);
@@ -377,7 +388,7 @@ class CRM_Core_I18n_Schema {
    *
    * @return array
    */
-  static function schemaStructureTables($version = NULL, $force = FALSE) {
+  public static function schemaStructureTables($version = NULL, $force = FALSE) {
     static $_tables = NULL;
     if ($_tables === NULL || $force) {
       if ($version) {
@@ -386,7 +397,7 @@ class CRM_Core_I18n_Schema {
         // class loader look for file like - CRM/Core/I18n/SchemaStructure/4/1/0.php which is not what we want to be loaded
         require_once "CRM/Core/I18n/SchemaStructure_{$latest}.php";
         $class = "CRM_Core_I18n_SchemaStructure_{$latest}";
-        $tables  =& $class::tables();
+        $tables =& $class::tables();
       }
       else {
         $tables = CRM_Core_I18n_SchemaStructure::tables();
@@ -401,7 +412,7 @@ class CRM_Core_I18n_Schema {
    *
    * @return mixed
    */
-  static function getLatestSchema($version) {
+  public static function getLatestSchema($version) {
     // remove any .upgrade sub-str from version. Makes it easy to do version_compare & give right result
     $version = str_ireplace(".upgrade", "", $version);
 
@@ -424,13 +435,17 @@ class CRM_Core_I18n_Schema {
   }
 
   /**
-   * CREATE INDEX queries for a given locale and table
+   * CREATE INDEX queries for a given locale and table.
    *
-   * @param $locale string  locale for which the queries should be created (null to create original indices)
-   * @param $table string   table for which the queries should be created
-   * @param $class string   schema structure class to use
+   * @param string $locale
+   *   locale for which the queries should be created (null to create original indices).
+   * @param string $table
+   *   table for which the queries should be created.
+   * @param string $class
+   *   schema structure class to use.
    *
-   * @return array          array of CREATE INDEX queries
+   * @return array
+   *   array of CREATE INDEX queries
    */
   private static function createIndexQueries($locale, $table, $class = 'CRM_Core_I18n_SchemaStructure') {
     $indices =& $class::indices();
@@ -463,18 +478,25 @@ class CRM_Core_I18n_Schema {
   }
 
   /**
-   * CREATE VIEW query for a given locale and table
+   * CREATE VIEW query for a given locale and table.
    *
-   * @param $locale string  locale of the view
-   * @param $table string   table of the view
-   * @param $dao object     a DAO object to run DESCRIBE queries
-   * @param $class string   schema structure class to use
-   *
-   * @return array          array of CREATE INDEX queries
+   * @param string $locale
+   *   locale of the view.
+   * @param string $table
+   *   table of the view.
+   * @param CRM_Core_DAO $dao
+   *   A DAO object to run DESCRIBE queries.
+   * @param string $class
+   *   schema structure class to use.
+   * @param bool $isUpgradeMode
+   *   Are we in upgrade mode therefore only build based off table not class
+   * @return array
+   *   array of CREATE INDEX queries
    */
-  private static function createViewQuery($locale, $table, &$dao, $class = 'CRM_Core_I18n_SchemaStructure') {
+  private static function createViewQuery($locale, $table, &$dao, $class = 'CRM_Core_I18n_SchemaStructure', $isUpgradeMode = FALSE) {
     $columns =& $class::columns();
     $cols = array();
+    $tableCols = array();
     $dao->query("DESCRIBE {$table}", FALSE);
     while ($dao->fetch()) {
       // view non-internationalized columns directly
@@ -483,10 +505,16 @@ class CRM_Core_I18n_Schema {
       ) {
         $cols[] = $dao->Field;
       }
+      $tableCols[] = $dao->Field;
     }
     // view intrernationalized columns through an alias
     foreach ($columns[$table] as $column => $_) {
-      $cols[] = "{$column}_{$locale} {$column}";
+      if (!$isUpgradeMode) {
+        $cols[] = "{$column}_{$locale} {$column}";
+      }
+      elseif (in_array("{$column}_{$locale}", $tableCols)) {
+        $cols[] = "{$column}_{$locale} {$column}";
+      }
     }
     return "CREATE OR REPLACE VIEW {$table}_{$locale} AS SELECT " . implode(', ', $cols) . " FROM {$table}";
   }
@@ -495,7 +523,7 @@ class CRM_Core_I18n_Schema {
    * @param $info
    * @param null $tableName
    */
-  static function triggerInfo(&$info, $tableName = NULL) {
+  public static function triggerInfo(&$info, $tableName = NULL) {
     // get the current supported locales
     $domain = new CRM_Core_DAO_Domain();
     $domain->find(TRUE);
@@ -542,7 +570,8 @@ class CRM_Core_I18n_Schema {
         foreach ($locales as $old) {
           $trigger[] = "ELSEIF NEW.{$column}_{$old} IS NOT NULL THEN";
           foreach (array_merge($locales, array(
-            $locale)) as $loc) {
+            $locale,
+          )) as $loc) {
             if ($loc == $old) {
               continue;
             }
@@ -553,7 +582,8 @@ class CRM_Core_I18n_Schema {
       }
 
       $sql = implode(' ', $trigger);
-      $info[] = array('table' => array($table),
+      $info[] = array(
+        'table' => array($table),
         'when' => 'BEFORE',
         'event' => array('UPDATE'),
         'sql' => $sql,
@@ -571,7 +601,8 @@ class CRM_Core_I18n_Schema {
         foreach ($locales as $old) {
           $trigger[] = "ELSEIF NEW.{$column}_{$old} IS NOT NULL THEN";
           foreach (array_merge($locales, array(
-            $locale)) as $loc) {
+            $locale,
+          )) as $loc) {
             if ($loc == $old) {
               continue;
             }
@@ -582,12 +613,13 @@ class CRM_Core_I18n_Schema {
       }
 
       $sql = implode(' ', $trigger);
-      $info[] = array('table' => array($table),
+      $info[] = array(
+        'table' => array($table),
         'when' => 'BEFORE',
         'event' => array('INSERT'),
         'sql' => $sql,
       );
     }
   }
-}
 
+}
